Adding DeriveSchemaUtils and abstract class DeriveSchema
