+++ maven-plugin/src/main/java/io/confluent/kafka/schemaregistry/maven/derive/schema/json/DeriveJsonSchema.java
@@ -18,23 +18,19 @@ package io.confluent.kafka.schemaregistry.maven.derive.schema.json;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.confluent.kafka.schemaregistry.maven.derive.schema.DeriveSchema;
 
 import java.util.List;
-import java.util.HashMap;
 import java.util.ArrayList;
-import java.util.Optional;
 
-import static io.confluent.kafka.schemaregistry.maven.derive.schema.DeriveSchema.mapper;
+import static io.confluent.kafka.schemaregistry.maven.derive.schema.DeriveSchemaUtils.getSortedKeys;
+import static io.confluent.kafka.schemaregistry.maven.derive.schema.DeriveSchemaUtils.mapper;
 import static io.confluent.kafka.schemaregistry.maven.derive.schema.json.MergeJsonUtils.mergeArrays;
 
-public class DeriveJsonSchema {
+public class DeriveJsonSchema extends DeriveSchema {
 
-  private static final HashMap<String, String> classToDataType = new HashMap<>();
-
-  static {
+  public DeriveJsonSchema() {
     // Map jackson node data type to type understood by json
     classToDataType.put(com.fasterxml.jackson.databind.node.DoubleNode.class.getName(), "number");
     classToDataType.put(com.fasterxml.jackson.databind.node.TextNode.class.getName(), "string");
@@ -48,43 +44,7 @@ public class DeriveJsonSchema {
     classToDataType.put(com.fasterxml.jackson.databind.node.MissingNode.class.getName(), "null");
   }
 
-  public static Optional<ObjectNode> getPrimitiveSchema(Object field)
-      throws JsonProcessingException {
-    // Generate Schema for Primitive type
-    String jsonInferredType = field.getClass().getName();
-    if (classToDataType.containsKey(jsonInferredType)) {
-      String schemaString = String.format("{\"type\" : \"%s\"}",
-          classToDataType.get(jsonInferredType));
-      return Optional.of(mapper.readValue(schemaString, ObjectNode.class));
-    }
-    return Optional.empty();
-  }
-
-  private static ArrayList<ObjectNode> getSchemaOfAllElements(List<JsonNode> messages, String name)
-      throws JsonProcessingException {
-    // Get schema of each message based on type
-    ArrayList<ObjectNode> schemaList = new ArrayList<>();
-    for (JsonNode message : messages) {
-      schemaList.add(getSchemaOfElement(message, name));
-    }
-    return schemaList;
-  }
-
-  private static ObjectNode getSchemaOfElement(JsonNode message, String name)
-      throws JsonProcessingException {
-    // Get schema of message based on type
-    Optional<ObjectNode> primitiveSchema = getPrimitiveSchema(message);
-    if (primitiveSchema.isPresent()) {
-      return primitiveSchema.get();
-    } else if (message instanceof ArrayNode) {
-      return getSchemaForArray(DeriveSchema.getListFromArray((ArrayNode) message), name);
-    } else {
-      ObjectNode objectNode = mapper.valueToTree(message);
-      return getSchemaForRecord(objectNode);
-    }
-  }
-
-  public static ObjectNode getSchemaForArray(List<JsonNode> messages, String name)
+  protected ObjectNode getSchemaForArray(List<JsonNode> messages, String name)
       throws JsonProcessingException {
     // Generate Schema for Array type
     ObjectNode schema = mapper.createObjectNode();
@@ -95,7 +55,7 @@ public class DeriveJsonSchema {
     return schema;
   }
 
-  public static ObjectNode getSchemaForRecord(ObjectNode message)
+  protected ObjectNode getSchemaForRecord(ObjectNode message)
       throws JsonProcessingException {
     // Generate Schema for Record type
     ObjectNode schema = mapper.createObjectNode();
@@ -103,7 +63,7 @@ public class DeriveJsonSchema {
     schema.set("properties", mapper.createObjectNode());
 
     // Loop over each field, get type of each field and insert into schema
-    for (String fieldName : DeriveSchema.getSortedKeys(message)) {
+    for (String fieldName : getSortedKeys(message)) {
       JsonNode field = message.get(fieldName);
       ObjectNode fields = (ObjectNode) schema.get("properties");
       fields.set(fieldName, getSchemaOfElement(field, fieldName));
@@ -111,7 +71,7 @@ public class DeriveJsonSchema {
     return schema;
   }
 
-  public static ObjectNode getSchemaForMultipleMessages(List<String> messages)
+  public ObjectNode getSchemaForMultipleMessages(List<String> messages)
       throws JsonProcessingException {
     /*
      Get schema for multiple messages. Exactly one schema is returned
@@ -121,8 +81,7 @@ public class DeriveJsonSchema {
     for (String message : messages) {
       messageObjects.add(mapper.readTree(message));
     }
-    ObjectNode schema = (ObjectNode) getSchemaForArray(
-        messageObjects, "").get("items");
+    JsonNode schema = getSchemaForArray(messageObjects, "").get("items");
     ObjectNode schemaInformation = mapper.createObjectNode();
     schemaInformation.set("schema", schema);
     return schemaInformation;
